pyproject.toml
[project]
name = "workflow"
version = "1.2.2.1"
readme = { file = "README.md", content-type = "text/markdown" }
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [
  { name = "Luis Fernando Umaña Castro", email = "luis.umanacastro@ucr.ac.cr" },
]
classifiers = ["Development Status :: 2 - Pre-Alpha"]
dependencies = ["appdirs", "bibtexparser", "click", "pyyaml"]

[project.scripts]
workflow = "main:cli"
cleta = "lectkit.cleta:cli"
crete = "lectkit.crete:cli"
inittex = "itep.create:cli"
relink = "itep.links:cli"
nofi = "lectkit.nofi:cli"

[build-system]
requires = ["uv-build"]
build-backend = "uv_build"

[tool.pytest.ini_options]
pythonpath = "."
